Change data types of context, phrase and clue sprint

+++ database/migrations/2019_07_16_084416_create_context_sprint.php
@@ -16,9 +16,9 @@ class CreateContextSprint extends Migration
         Schema::connection('pgsql')->create('context_sprint', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('topic_id');
-            $table->unsignedInteger('context_id');
-            $table->unsignedInteger('solution_phrase_id');
-            $table->unsignedInteger('wrong_phrase_id');
+            $table->unsignedBigInteger('context_id');
+            $table->unsignedBigInteger('solution_phrase_id');
+            $table->unsignedBigInteger('wrong_phrase_id');
             $table->integer('bucket');
 
             $table->foreign('topic_id')->references('id')->on('context_topics')->onUpdate('cascade')->onDelete('cascade');
